requirements.txt
langchain_community
langchain_core
langchain_groq
langchain_openai
streamlit
pypdf
PyPDF2
python-dotenv
groq
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
cassio
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
langchain-objectbox
langchain-astradb